feat(playground): add login endpoint
