+++ playground/server/api/built-in-password.ts
@@ -1,0 +1,18 @@
+export default eventHandler(async (event) => {
+  const { password } = await readBody(event)
+
+  if (password !== '123456') {
+    throw createError({
+      statusCode: 401,
+      message: 'Wrong password',
+    })
+  }
+  await setUserSession(event, {
+    user: {
+      password: 'admin',
+    },
+    loggedInAt: Date.now(),
+  })
+
+  return {}
+})
